requirements.txt
flask==2.3.3
werkzeug==2.3.7
face-recognition==1.3.0
Pillow==8.3.1
numpy==1.21.0
python-dotenv==0.19.0
groq==0.4.2 